fix: handle events deletion when API has no event
https://gravitee.atlassian.net/browse/APIM-3625

+++ gravitee-apim-repository/gravitee-apim-repository-jdbc/src/main/java/io/gravitee/repository/jdbc/management/JdbcEventRepository.java
@@ -307,6 +307,10 @@ public class JdbcEventRepository extends JdbcAbstractPageableRepository<Event> i
                 apiId
             );
 
+            if (eventToDelete.isEmpty()) {
+                return 0;
+            }
+
             String propertiesDeleteQuery =
                 "delete from " + EVENT_PROPERTIES + " where event_id in (" + getOrm().buildInClause(eventToDelete) + ")";
             jdbcTemplate.update(
